requirements.txt
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pygame==2.5.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
setuptools==68.2.2
six==1.16.0
tqdm==4.66.1
tzdata==2023.4
wheel==0.41.2
wordcloud==1.9.3
